requirements.txt
xai-sdk>=0.1.0
jinja2>=3.1.0
python-dotenv>=1.0.0
